Move aggregation logic to new AggregationUtil
