+++ src/utils/basal.js
@@ -287,61 +287,3 @@ export const countDistinctSuspends = (data) => {
 
   return returnData;
 };
-
-/**
- * postProcessBasalAggregations
- *
- * Post processor for crossfilter reductio basal aggregations
- *
- * @param {Function} priorResults - returns the data from the active crossfilter reductio reducer
- * @returns {Object} formatted total and subtotal data for basal aggregations
- */
-export const postProcessBasalAggregations = priorResults => () => {
-  const data = _.filter(
-    _.cloneDeep(priorResults()),
-    ({ value: { dataList } }) => !_.isEmpty(dataList)
-  );
-
-  const processedData = {};
-
-  _.each(data, dataForDay => {
-    const {
-      value: {
-        dataList,
-        suspend, // TODO: countDistinctSuspends()
-        temp,
-      },
-    } = dataForDay;
-
-    processedData[dataForDay.key] = {
-      data: dataList,
-      total: _.reduce([suspend, temp], (acc, { count = 0 }) => acc + count, 0),
-      subtotals: {
-        suspend: suspend.count,
-        temp: temp.count,
-      },
-    };
-
-    _.assign(
-      processedData[dataForDay.key],
-      countAutomatedBasalEvents(processedData[dataForDay.key])
-    );
-
-    // No need to return the data - we only want the aggregations
-    delete processedData[dataForDay.key].data;
-  });
-
-  return {
-    summary: {
-      total: _.sumBy(_.values(processedData), dateData => dateData.total),
-      subtotals: _.reduce(_.map(_.values(processedData), 'subtotals'), (acc, subtotals) => {
-        const tags = _.keysIn(subtotals);
-        _.each(tags, tag => {
-          acc[tag] = (acc[tag] || 0) + subtotals[tag];
-        });
-        return acc;
-      }, {}),
-    },
-    byDate: processedData,
-  };
-};
