+++ src/utils/bolus.js
@@ -353,65 +353,3 @@ export function getAnnotations(insulinEvent) {
   const annotations = _.get(bolus, 'annotations', []);
   return annotations;
 }
-
-/**
- * postProcessBolusAggregations
- *
- * Post processor for crossfilter reductio bolus aggregations
- *
- * @param {Function} priorResults - returns the data from the active crossfilter reductio reducer
- * @returns {Object} formatted total and subtotal data for bolus aggregations
- */
-export const postProcessBolusAggregations = priorResults => () => {
-  const data = _.filter(
-    _.cloneDeep(priorResults()),
-    ({ value: { dataList } }) => !_.isEmpty(dataList)
-  );
-
-  const processedData = {};
-
-  _.each(data, dataForDay => {
-    const {
-      value: {
-        correction,
-        dataList,
-        extended,
-        interrupted,
-        manual,
-        override,
-        underride,
-        wizard,
-      },
-    } = dataForDay;
-
-    processedData[dataForDay.key] = {
-      total: dataList.length,
-      subtotals: {
-        correction: correction.count,
-        extended: extended.count,
-        interrupted: interrupted.count,
-        manual: manual.count,
-        override: override.count,
-        underride: underride.count,
-        wizard: wizard.count,
-      },
-    };
-
-    // No need to return the data - we only want the aggregations
-    delete processedData[dataForDay.key].data;
-  });
-
-  return {
-    summary: {
-      total: _.sumBy(_.values(processedData), dateData => dateData.total),
-      subtotals: _.reduce(_.map(_.values(processedData), 'subtotals'), (acc, subtotals) => {
-        const tags = _.keysIn(subtotals);
-        _.each(tags, tag => {
-          acc[tag] = (acc[tag] || 0) + subtotals[tag];
-        });
-        return acc;
-      }, {}),
-    },
-    byDate: processedData,
-  };
-};
