+++ src/utils/device.js
@@ -47,49 +47,3 @@ export function getPumpVocabulary(manufacturer) {
     vocabulary.default
   );
 }
-
-/**
- * postProcessBolusAggregations
- *
- * Post processor for crossfilter reductio bolus aggregations
- *
- * @param {Function} priorResults - returns the data from the active crossfilter reductio reducer
- * @returns {Object} formatted total and subtotal data for bolus aggregations
- */
-export const postProcessCalibrationAggregations = priorResults => () => {
-  const data = _.filter(
-    _.cloneDeep(priorResults()),
-    ({ value: { dataList } }) => !_.isEmpty(dataList)
-  );
-
-  const processedData = {};
-
-  _.each(data, dataForDay => {
-    const {
-      value: {
-        calibration,
-      },
-    } = dataForDay;
-
-    processedData[dataForDay.key] = {
-      total: calibration.count,
-      subtotals: {
-        calibration: calibration.count,
-      },
-    };
-  });
-
-  return {
-    summary: {
-      total: _.sumBy(_.values(processedData), dateData => dateData.total),
-      subtotals: _.reduce(_.map(_.values(processedData), 'subtotals'), (acc, subtotals) => {
-        const tags = _.keysIn(subtotals);
-        _.each(tags, tag => {
-          acc[tag] = (acc[tag] || 0) + subtotals[tag];
-        });
-        return acc;
-      }, {}),
-    },
-    byDate: processedData,
-  };
-};
